+++ src/main/java/at/porscheinformatik/seleniumcomponents/clarity/ClarityMultiSelectComboboxContainerComponent.java
@@ -2,11 +2,13 @@ package at.porscheinformatik.seleniumcomponents.clarity;
 
 import static at.porscheinformatik.seleniumcomponents.WebElementSelector.*;
 
+import java.util.function.Predicate;
+
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponentFactory;
+import at.porscheinformatik.seleniumcomponents.SeleniumComponentListFactory;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 import at.porscheinformatik.seleniumcomponents.clarity.ClarityComboboxOptionsComponent.AbstractClarityComboboxOptionComponent;
-import at.porscheinformatik.seleniumcomponents.clarity.ClarityMultiSelectComboboxComponent.AbstractClarityComboboxLabelPillComponent;
 
 /**
  * @author TechScar
@@ -17,7 +19,13 @@ public class ClarityMultiSelectComboboxContainerComponent<OPTION_TYPE extends Ab
 {
     public final ClarityMultiSelectComboboxComponent<OPTION_TYPE, PILL_TYPE> multiSelect;
 
-    // ---
+    public static ClarityMultiSelectComboboxContainerComponent<DefaultClarityComboboxOptionComponent,
+        DefaultClarityComboboxLabelPillComponent> within(
+        SeleniumComponent parent)
+    {
+        return new ClarityMultiSelectComboboxContainerComponent<>(parent, selectByTagName("clr-combobox-container"),
+            DefaultClarityComboboxOptionComponent::new, DefaultClarityComboboxLabelPillComponent::new);
+    }
 
     public static <OPTION_TYPE extends AbstractClarityComboboxOptionComponent,
         PILL_TYPE extends AbstractClarityComboboxLabelPillComponent> ClarityMultiSelectComboboxContainerComponent<OPTION_TYPE, PILL_TYPE> within(
@@ -26,6 +34,14 @@ public class ClarityMultiSelectComboboxContainerComponent<OPTION_TYPE extends Ab
     {
         return new ClarityMultiSelectComboboxContainerComponent<>(parent, selectByTagName("clr-combobox-container"),
             optionFactory, pillFactory);
+    }
+
+    public static ClarityMultiSelectComboboxContainerComponent<DefaultClarityComboboxOptionComponent,
+        DefaultClarityComboboxLabelPillComponent> byTestId(
+        SeleniumComponent parent, String testId)
+    {
+        return new ClarityMultiSelectComboboxContainerComponent<>(parent, selectByTestId(testId),
+            DefaultClarityComboboxOptionComponent::new, DefaultClarityComboboxLabelPillComponent::new);
     }
 
     public static <OPTION_TYPE extends AbstractClarityComboboxOptionComponent,
@@ -51,13 +67,41 @@ public class ClarityMultiSelectComboboxContainerComponent<OPTION_TYPE extends Ab
             optionFactory, pillFactory);
     }
 
-    // ---
-
     public ClarityMultiSelectComboboxContainerComponent(SeleniumComponent parent, WebElementSelector selector,
         SeleniumComponentFactory<OPTION_TYPE> optionFactory, SeleniumComponentFactory<PILL_TYPE> pillFactory)
     {
         super(parent, selector);
 
         multiSelect = new ClarityMultiSelectComboboxComponent<>(this, optionFactory, pillFactory);
+    }
+
+    public void clear()
+    {
+        multiSelect.clear();
+    }
+
+    public PILL_TYPE findLabelPill(Predicate<PILL_TYPE> predicate)
+    {
+        return multiSelect.findLabelPill(predicate);
+    }
+
+    public PILL_TYPE getPillByLabel(String partialText)
+    {
+        return multiSelect.getPillByLabel(partialText);
+    }
+
+    public void assertSelected(String partialText)
+    {
+        multiSelect.assertSelected(partialText);
+    }
+
+    public SeleniumComponentListFactory<PILL_TYPE> getLabelPills()
+    {
+        return multiSelect.getLabelPills();
+    }
+
+    public void removeByLabel(String partialText)
+    {
+        multiSelect.removeByLabel(partialText);
     }
 }